requirements.txt
aniso8601==8.0.0
argcomplete==1.10.0
beautifulsoup4==4.8.0
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.14.56
botocore==1.17.56
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cymem==2.0.3
Cython==0.29.21
decorator==4.4.2
dnspython==2.0.0
docutils==0.15.2
docx2txt==0.8
dominate==2.5.2
EbookLib==0.17.1
email-validator==1.1.1
extract-msg==0.23.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.9
future==0.18.2
gensim==3.8.3
geocoder==1.38.1
graphpipe==1.0.4
gunicorn==20.0.4
idna==2.6
IMAPClient==2.1.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
murmurhash==1.0.2
newsapi==0.1.1
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
pdfminer.six==20181108
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pycryptodome==3.9.8
pygeocoder==1.2.5
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.1
PyYAML==5.3.1
ratelim==0.1.6
regex==2020.7.14
requests==2.18.4
requests-file==1.5.1
requests-oauthlib==1.3.0
reverse-geocoder==1.5.1
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.12.0
sklearn==0.0
smart-open==2.1.1
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==2.3.2
SpeechRecognition==3.8.1
SQLAlchemy==1.3.19
srsly==1.0.2
textract==1.6.3
thinc==7.4.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.48.2
tweepy==3.9.0
tzlocal==1.5.1
urllib3==1.22
vaderSentiment==3.3.2
visitor==0.1.3
wasabi==0.8.0
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.3.3
yweather==0.1.1
./en_core_web_md-2.3.1.tar.gz
./en_core_web_sm-2.3.1.tar.gz